Remove an extra User.findOne for user.* API routes
